move account public key to keys
