+++ flowkit/tests/testnet_test.go
@@ -67,7 +67,7 @@ func initTestnet(t *testing.T) (gateway.Gateway, *flowkit.State, flowkit.Service
 	flowAccount, _, err := flow.CreateAccount(
 		context.Background(),
 		funder,
-		[]flowkit.AccountPublicKey{{
+		[]accounts2.AccountPublicKey{{
 			Public:   testKey.PublicKey(),
 			Weight:   1000,
 			SigAlgo:  crypto.ECDSA_P256,
