+++ internal/accounts/accounts_test.go
@@ -165,7 +165,7 @@ func Test_Create(t *testing.T) {
 
 		srv.CreateAccount.Run(func(args mock.Arguments) {
 			acc := args.Get(1).(*accounts.Account)
-			keys := args.Get(2).([]flowkit.AccountPublicKey)
+			keys := args.Get(2).([]accounts.AccountPublicKey)
 			assert.Equal(t, "emulator-account", acc.Name)
 			assert.Len(t, keys, 1)
 			assert.Equal(t, fmt.Sprintf("0x%s", pkey), keys[0].Public.String())
@@ -188,7 +188,7 @@ func Test_Create(t *testing.T) {
 
 		srv.CreateAccount.Run(func(args mock.Arguments) {
 			acc := args.Get(1).(*accounts.Account)
-			keys := args.Get(2).([]flowkit.AccountPublicKey)
+			keys := args.Get(2).([]accounts.AccountPublicKey)
 			assert.Equal(t, "emulator-account", acc.Name)
 			assert.Len(t, keys, 2)
